Add app logic
