+++ CMakeLists.txt
@@ -7,7 +7,8 @@ set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 
 add_subdirectory(src)
+add_subdirectory(app)
 
-add_executable(run_example example.cpp)
-target_include_directories(run_example PUBLIC src)
-target_link_libraries(run_example PUBLIC 3d_renderer)
+add_executable(run main.cpp)
+target_include_directories(run PUBLIC src app)
+target_link_libraries(run PUBLIC 3d_renderer app)
