+++ app/application.h
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace app {
+
+class Application {
+public:
+    void Run();
+};
+
+}  // namespace app
