+++ app/engine_fwd.h
@@ -1,0 +1,5 @@
+#include "camera.h"
+#include "object3d.h"
+#include "reference.h"
+#include "renderer.h"
+#include "scene.h"
