+++ app/except.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace app::except {
+
+void React();
+
+};  // namespace app::except
