+++ main.cpp
@@ -1,0 +1,11 @@
+#include "application.h"
+#include "except.h"
+
+int main() {
+    try {
+        app::Application app;
+        app.Run();
+    } catch (...) {
+        app::except::React();
+    }
+}
